pyproject.toml
[tool.poetry]
name = "whishlist"
version = "0.1.0"
description = ""
authors = ["Rhama Krisner <rhama.krisner@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
